Revert "Merge branch 'v1.8.x' into master (#1706)"
This reverts commit 66e24bf8ec271538d459fc4efdb6b48d64fcb7e5.

+++ cmd/cloud-server/logics/load-balancer/validate.go
@@ -60,7 +60,7 @@ func validateInstType(instType enumor.InstType) error {
 	return nil
 }
 
-func validateWeight(weightPtr *int64) error {
+func validateWeight(weightPtr *int) error {
 	if weightPtr == nil {
 		return errors.New("invalid weight: nil")
 	}
